requirements.txt
Django==2.0.6
djangorestframework==3.8.2
numpy==1.14.5
pytz==2018.4
scipy==1.1.0
